CDH-76213: Use int64_t for ScanRange::bytes_read_
When scanning an unsplittable file (e.g. gzipped text), the
ScanRange's size is the whole file size. If the file is larger
than 2GB, ScanRange::bytes_read_ can overflow its 32-bit
integer value, leading to a failed query.

This changes ScanRange::bytes_read_ to an int64_t.

Testing:
 - Verified by hand with a 2+GB gzipped text file that this
   succeeds where it previous failed.

Reviewed-on: https://gerrit.sjc.cloudera.com/40159
Tested-by: Joseph McDonnell <joemcdonnell@cloudera.com>

+++ be/src/runtime/io/request-ranges.h
@@ -376,7 +376,7 @@ class ScanRange : public RequestRange {
   boost::mutex lock_;
 
   /// Number of bytes read so far for this scan range
-  int bytes_read_;
+  int64_t bytes_read_;
 
   /// Status for this range. This is non-ok if is_cancelled_ is true.
   /// Note: an individual range can fail without the RequestContext being
